Scrappy change that provides support for Option<T> fields.
The compiled path supports Option<T> via the existing 'Collection' mechanism (to be renamed to something like 'Container').
Still not 100% sure about this choice, but I tried several paths and this seemed the best.

Left a bunch of the false paths in the code here for future reference.
Fast follow up to clean this up now.