Add basic presentation to the cli.
This change focuses on:
1. Presenting error information to the user interface.
2. Presenting basic usage information to the user interface.
3. Setup for using 'nargs' in the public Api.

1: We present human readable descriptions for the various error conditions.
Additionally, the output gives an indication of which token from the input caused the error.

2: We present the single line 'usage: PROGRAM ..' description.
The help flag is always presented first, then the cli options (in alphabetical order), and
finally the cli arguments (in positional order).
When a short name is provided for the option, it is used to minimize character usage.

3. The original Api concept was to always derive 'nargs' from the field + generic.
In retrospect, this will require many separate but similar implementations for container structs
(ex: regular `Vec`, `AtLeastOneVec`, `VecPair`, `VecTriple`, etc).
Thefore, part of this change starts to pave the way to expose 'nargs' explicitly over the Api, when appropriate.