Add sub-command functionality.
*Warning* intentionally hacky change.
This is a proto-type that supports sub-commands.
Done with a mixture of composition and "parallel implementation".
Long term plan is to maintain roughly this structure, but with
lots of code cleanup, renaming, and possibly 1 less "parallel implementation".

Sub-command functionality works by switching the argument parser into sub-command mode.
From here, the CLI developer adds parameters like normal, except each is flagged with
the sub-command instance it belongs to.
Under the hood, we build out a 'root' parser and 'branch' parsers, and then invoke these
in sequence based off the runtime.

The switching type of the sub-command requires `std::fmt::Display`.
Additionally, it implicitly requires that `std::fmt::Display` is the inverse
to `std::str::FromStr`, and visa-versa.
Using a type that breaks this requirement may result in undefined behaviour.
For example, the switching type may correctly capture the CLI input, but the associated
sub-command will not be found.