Change sub-command setup to specify a single variant at a time.
It was a little awkward to specify sub-commands via `SubCommandParser::add(..)`,
since multiple parameters on the sub-command would each need a separate call.
This change uses a simpler mechanism, where the entire sub-command
is added in one shot via `SubCommandParser::command(..)`.
This change is backwards compatible, but support for `SubCommandParser::add(..)`
will be removed in the future.

Included several documentation updates as well.