Support choices documentation in the derive Api.
This change adds a new attribute '#[blarg(choices)]' to derive Api parameters.
When used without assignment (as above), it wires the choices via a method on the parameter's type called 'blarg_choices'.
When used with assignment (as in '#[blarg(choices = my_func)]'), it wires the choices via the provided 'my_func' function.

In order to support the common choices use-case, we also add a new derive macro 'BlargChoices' which applies to enums.
The change includes other minor refactoring, including:
* Moving public traits to 'blarg::prelude'.
* Moving derive macros to 'blarg::derive'.