Add parameter type hints.
This change adds type documentation on parameters, when applicable.
If there is choices documentation, we don't add the type hint as it is superfluous.
We also take care to only document the type when useful - for example, the type hint
is not informative on a boolean switch.

With this change, the help message documentation has become quite dense.
Will look to provide a follow-up UX improvement.