Address a few usability concerns.
* Introduce fallible `Collectable`s.
This allows for collection implementations that fail at capture time.
We change `HashSet`'s implemention to be one such fallible `Collectable`.
* Protect against the condition invariant at parser build time.
This allows us to detect the violation earlier, and cleans up some internal state.
* Change parser `build` to directly print and exit upon invalid configuration.
In case users need an option without sys exit, the method `build_parser` exists.
This follows a similar pattern to parsing itself with `parse` and `parse_tokens`.
* Change parse error to distinguish where the error occurred.
This is a minor UX touch-up that follows from the documentation (we document the
match and capture phases distinctly).
* More documentation touch-ups.

This is a breaking change, but given how unlikely it is anyone is using 1.0.0, we've
simply bumped the version to 1.0.1 (and will yank 1.0.0).