Improve rendering width heuristic and other touch-ups.
Previously, the middle column width would default to the minimum size if the non-middle (left/right) were too large.
This would happen even when a middle greater than the minimum was available in the terminal.
This change re-works the middle column width heuristic to make a better choice about this.

Additionally, includes a few other changes:
* To help debug cases like the above, introduced a "debug" feature which uses the tracing library.
* Another previous change (c399f930) introduced a panic if the sub-command respected the condition invariant, but wasn't a valid sub-command.
* Updates to documentation and examples.